+++ src/backend/cuda/include/yaksuri_cudai.h
@@ -21,12 +21,7 @@ extern "C" {
 #endif
 /* *INDENT-ON* */
 
-#define YAKSURI_CUDAI_CUDA_ERR_CHECK(cerr)                              \
-    do {                                                                \
-        if (cerr != cudaSuccess) {                                      \
-            fprintf(stderr, "CUDA Error (%s:%s,%d): %s\n", __func__, __FILE__, __LINE__, cudaGetErrorString(cerr)); \
-        }                                                               \
-    } while (0)
+#include <yaksuri_cudai_base.h>
 
 #define YAKSURI_CUDAI_CUDA_ERR_CHKANDJUMP(cerr, rc, fn_fail)            \
     do {                                                                \
@@ -36,47 +31,6 @@ extern "C" {
             goto fn_fail;                                               \
         }                                                               \
     } while (0)
-
-typedef struct {
-    int ndevices;
-    cudaStream_t *stream;
-    bool **p2p;
-} yaksuri_cudai_global_s;
-extern yaksuri_cudai_global_s yaksuri_cudai_global;
-
-typedef struct yaksuri_cudai_md_s {
-    union {
-        struct {
-            int count;
-            intptr_t stride;
-            struct yaksuri_cudai_md_s *child;
-        } contig;
-        struct {
-            struct yaksuri_cudai_md_s *child;
-        } resized;
-        struct {
-            int count;
-            int blocklength;
-            intptr_t stride;
-            struct yaksuri_cudai_md_s *child;
-        } hvector;
-        struct {
-            int count;
-            int blocklength;
-            intptr_t *array_of_displs;
-            struct yaksuri_cudai_md_s *child;
-        } blkhindx;
-        struct {
-            int count;
-            int *array_of_blocklengths;
-            intptr_t *array_of_displs;
-            struct yaksuri_cudai_md_s *child;
-        } hindexed;
-    } u;
-
-    uintptr_t extent;
-    uintptr_t num_elements;
-} yaksuri_cudai_md_s;
 
 typedef struct yaksuri_cudai_type_s {
     void (*pack) (const void *inbuf, void *outbuf, uintptr_t count, yaksuri_cudai_md_s * md,
